requirements.txt
# Telegram Bot
aiogram>=3.7.0

# AI
google-generativeai>=0.8.0
openai>=1.0.0

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
greenlet>=3.0.0

# Admin Panel
sqladmin>=0.19.0
jinja2>=3.1.0

# HTTP Client (for parsing)
httpx>=0.26.0
beautifulsoup4>=4.12.0

# Telegram Markdown
telegramify-markdown>=0.5.4

# Settings
pydantic-settings>=2.1.0

# Session middleware
itsdangerous>=2.1.0
